Moved time functions to separate file and added formatDate function
